+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed:
+- Navigate back icon on task page
+- Unordered list color in editor
+- Editor menu background color
+
+### Changed:
+- Improved task input fields layout
+- Improved chip layout
+- Tweak spacing in journal filters
+
+## [0.8.264] - 2023-02-11
+### Fixed:
 - Audio playback restart on navigate to audio entry
 
 ### Changed: